Refactor browser_stack_devices.rb
