+++ bin/maze-runner
@@ -10,7 +10,6 @@ require_relative '../lib/maze'
 require_relative '../lib/maze/appium_server'
 require_relative '../lib/maze/bitbar_utils'
 require_relative '../lib/maze/bitbar_devices'
-require_relative '../lib/maze/browser_stack_devices'
 require_relative '../lib/maze/browser_stack_utils'
 require_relative '../lib/maze/bugsnag_config'
 require_relative '../lib/maze/capabilities'
@@ -18,6 +17,7 @@ require_relative '../lib/maze/compare'
 require_relative '../lib/maze/docker'
 require_relative '../lib/maze/document_server'
 require_relative '../lib/maze/errors'
+require_relative '../lib/maze/farm/browser_stack/appium/devices'
 require_relative '../lib/maze/helper'
 require_relative '../lib/maze/logger'
 require_relative '../lib/maze/macos_utils'
@@ -108,11 +108,11 @@ class MazeRunnerEntry
     if options[Maze::Option::LIST_DEVICES]
       case options[Maze::Option::FARM]
       when :bs
-        Maze::BrowserStackDevices.list_devices('ios')
-        Maze::BrowserStackDevices.list_devices('android')
+        Maze::Farm::BrowserStack::Devices.list_devices('ios')
+        Maze::Farm::BrowserStack::Devices.list_devices('android')
       else
-        Maze::BrowserStackDevices.list_devices('ios')
-        Maze::BrowserStackDevices.list_devices('android')
+        Maze::Farm::BrowserStack::Devices.list_devices('ios')
+        Maze::Farm::BrowserStack::Devices.list_devices('android')
       end
       exit 0
     end
